+++ collection_helper_test.go
@@ -4,7 +4,11 @@
 
 package collection_test
 
-import "fmt"
+import (
+	"fmt"
+
+	"github.com/bborbe/collection"
+)
 
 type User struct {
 	Firstname string
@@ -20,3 +24,9 @@ func (u User) Equal(user User) bool {
 func (u User) HashCode() string {
 	return fmt.Sprintf("%#v", u)
 }
+
+// CustomStringType is a string-based type to test Set UnmarshalText/MarshalText
+type CustomStringType string
+
+// CustomStringTypeSet is a type alias for collection.Set to test that type aliases work
+type CustomStringTypeSet = collection.Set[CustomStringType]